money long타입으로 수정

+++ Assets/Scripts/Application Layer/UnitSystem/Interface/IInventory.cs
@@ -7,6 +7,6 @@ public interface IInventory
     IReadOnlyList<IInventorySlot> inventorySlots { get; }
     int currentSlotCnt { get; }
     Transform GetTransform();
-    int money { get; }
+    long money { get; }
     int carrot { get; }
 }
